I solved some errors and put all functions in functions.js, created function to start a new game and changes on css and html.
I misunderstood solitaire spider and I thought cards were
completely random and actually this game uses 2 deck of
cards, so I created arrays representing these 2 decks.

I moved the menu to the top and styled the .slots-section
to be on the center

Also solved a problem with verifyIfDraggable() not
working.

And I created newGame(function) which takes all cards
of cards1 array and two of cards2.